+++ src/shun.c
@@ -49,6 +49,7 @@ void shun_page(void){
   const char *zShun = P("shun");
   const char *zAccept = P("accept");
   const char *zRcvid = P("rcvid");
+  int reviewList = P("review")!=0;
   int nRcvid = 0;
   int numRows = 3;
   char *zCanonical = 0;
@@ -87,7 +88,7 @@ void shun_page(void){
     p = zCanonical;
     while( *p ){
       int nUuid = strlen(p);
-      if( !hname_validate(p, nUuid) ){
+      if( !(reviewList || hname_validate(p, nUuid)) ){
         @ <p class="generalError">Error: Bad artifact IDs.</p>
         fossil_free(zCanonical);
         zCanonical = 0;
@@ -100,10 +101,9 @@ void shun_page(void){
     zUuid = zCanonical;
   }
   style_header("Shunned Artifacts");
-  if( zUuid && P("sub") ){
+  if( zUuid && P("sub") && cgi_csrf_safe(2) ){
     const char *p = zUuid;
     int allExist = 1;
-    login_verify_csrf_secret();
     while( *p ){
       db_multi_exec("DELETE FROM shun WHERE uuid=%Q", p);
       if( !db_exists("SELECT 1 FROM blob WHERE uuid=%Q", p) ){
@@ -129,10 +129,9 @@ void shun_page(void){
       @ can pulled in from other repositories.</p>
     }
   }
-  if( zUuid && P("add") ){
+  if( zUuid && P("add") && cgi_csrf_safe(2) ){
     const char *p = zUuid;
     int rid, tagid;
-    login_verify_csrf_secret();
     while( *p ){
       db_multi_exec(
         "INSERT OR IGNORE INTO shun(uuid,mtime)"
@@ -158,6 +157,59 @@ void shun_page(void){
     @ have been shunned.  They will no longer be pushed.
     @ They will be removed from the repository the next time the repository
     @ is rebuilt using the <b>fossil rebuild</b> command-line</p>
+  }
+  if( zUuid && reviewList ){
+    const char *p;
+    int nTotal = 0;
+    int nOk = 0;
+    @ <table class="shun-review"><tbody><tr><td>
+    for( p = zUuid ; *p ; p += strlen(p)+1 ){
+      int rid = symbolic_name_to_rid(p, 0);
+      nTotal++;
+      if( rid < 0 ){
+        @ Ambiguous<br>
+      }else if( rid == 0 ){
+        if( !hname_validate(p, strlen(p)) ){
+          @ Bad artifact<br>
+        }else if(db_int(0, "SELECT 1 FROM shun WHERE uuid=%Q", p)){
+          @ Already shunned<br>
+        }else{
+          @ Unknown<br>
+        }
+      }else{
+        char *zCmpUuid = db_text(0,
+            "SELECT uuid"
+            "  FROM blob, rcvfrom"
+            " WHERE rid=%d"
+            "   AND rcvfrom.rcvid=blob.rcvid",
+            rid);
+        if( fossil_strcmp(p, zCmpUuid)==0 ){
+          nOk++;
+          @ OK</br>
+        }else{
+          @ Abbreviated<br>
+        }
+      }
+    }
+    @ </td><td>
+    for( p = zUuid ; *p ; p += strlen(p)+1 ){
+      int rid = symbolic_name_to_rid(p, 0);
+      if( rid > 0 ){
+        @ <a href="%R/artifact/%s(p)">%s(p)</a><br>
+      }else{
+        @ %s(p)<br>
+      }
+    }
+    @ </td></tr></tbody></table>
+    @ <p class="shunned">
+    if( nOk < nTotal){
+      @ <b>Warning:</b> Not all artifacts
+    }else if( nTotal==1 ){
+      @ The artifact is present and
+    }else{
+      @ All %i(nOk) artifacts are present and
+    }
+    @ can be shunned with its hash above.</p>
   }
   if( zRcvid ){
     nRcvid = atoi(zRcvid);
@@ -200,9 +252,15 @@ void shun_page(void){
       }
       db_finalize(&q);
     }
+  }else if( zUuid && reviewList ){
+    const char *p;
+    for( p = zUuid ; *p ; p += strlen(p)+1 ){
+      @ %s(p)
+    }
   }
   @ </textarea>
   @ <input type="submit" name="add" value="Shun">
+  @ <input type="submit" name="review" value="Review">
   @ </div></form>
   @ </blockquote>
   @
